canton: split the disclosure service into acs, server, and flows files and drop the daCoin placeholder
